further optimisation with a vectorise layer
accuracy increases on the last commit.

evaluate can be called with a singular tensor database the targets (y) will be obtained from x

this method allows passing in raw strings and the vectorisation will handle the rest.

the model can only be exported as a TF and not a keras archive
to preserve the preprocesing layer which can be leveraged by the
consumer chatbot module
i.e, the chatbot will no longer need to do text preprocessing as well.